# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of AquaTrack\n"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://aquatrack-backend-bmxm.onrender.com"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Operations about users."
    },
    {
      "name": "Water consumption",
      "description": "Operations about water consumption."
    }
  ],
  "paths": {
    "/users/signup": {
      "post": {
        "tags": ["Users"],
        "summary": "Sign up a user",
        "operationId": "SignupUser",
        "description": "Sign up a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": 1234567890
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sign up a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered a user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/signin": {
      "post": {
        "tags": ["Users"],
        "summary": "Sign in a user",
        "operationId": "SigninUser",
        "description": "Sign in a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": 1234567890
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign in a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in an user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": ["Users"],
        "summary": "Logout a user",
        "operationId": "LogoutUser",
        "description": "Logout a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logout a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "User was logout successfully!"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/refresh": {
      "post": {
        "tags": ["Users"],
        "summary": "Refresh token",
        "operationId": "refreshToken",
        "description": "Refresh token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "jhsdfetdtfgdhjkfbdsj"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully refreshed a session!"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/current": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user",
        "operationId": "getCurrentUser",
        "description": "Get current user ",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Current user information!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/users/update/{id}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update user",
        "operationId": "updateUser",
        "description": "Update a user with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "gender": {
                    "description": "User's gender",
                    "type": "string",
                    "enum": ["Woman", "Man"],
                    "example": "Woman"
                  },
                  "name": {
                    "description": "User's name",
                    "type": "string",
                    "example": "Jane Doe"
                  },
                  "weight": {
                    "description": "User's weight in kg",
                    "type": "number",
                    "example": 70
                  },
                  "activeTime": {
                    "description": "User's active time in hours",
                    "type": "number",
                    "example": 2
                  },
                  "amountOfWater": {
                    "description": "User's daily water consumption in liters",
                    "type": "number",
                    "example": 2.5
                  },
                  "photo": {
                    "description": "User's photo",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dzh0r2zb0/image/upload/v1721761102/hkxh93evnwllhlfienhj.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updates a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/water": {
      "post": {
        "tags": ["Water consumption"],
        "summary": "Create info about water consumption",
        "operationId": "createWaterConsumption",
        "description": "Create a water with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["userId", "date", "amount"],
                "properties": {
                  "userId": {
                    "description": "User's id",
                    "type": "string",
                    "example": "65e4decdd286b30065d54af8"
                  },
                  "date": {
                    "description": "Date of water consumption",
                    "type": "string",
                    "example": "2023-07-20T20:30"
                  },
                  "amount": {
                    "description": "Amount of water consumed",
                    "type": "number",
                    "example": 1.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a water consumption",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Water consumption record added successfully!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/water/{id}": {
      "put": {
        "tags": ["Water consumption"],
        "summary": "Update water consumption",
        "operationId": "updateWaterConsumption",
        "description": "Update a water consumption with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "description": "Date of water consumption",
                    "type": "string",
                    "example": "2023-07-20T20:30"
                  },
                  "amount": {
                    "description": "Amount of water consumed",
                    "type": "number",
                    "example": 1.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updates a water consumption",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Water consumption record updated successfully!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "delete": {
        "tags": ["Water consumption"],
        "summary": "Delete the entered amount of water",
        "operationId": "deleteWater",
        "description": "Delete the entered amount of water",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6693b609d6192024db757e57"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a water consumption",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Water consumption record deleted successfully!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": ["_id", "email", "password"],
        "properties": {
          "_id": {
            "description": "User's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "gender": {
            "description": "User's gender",
            "type": "string",
            "enum": ["Woman", "Man"],
            "example": "Woman"
          },
          "name": {
            "description": "User's name",
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "description": "User's email",
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "password": {
            "description": "User's password",
            "type": "string",
            "example": "1234567890"
          },
          "weight": {
            "description": "User's weight in kg",
            "type": "number",
            "example": 70
          },
          "activeTime": {
            "description": "User's active time in hours",
            "type": "number",
            "example": 2
          },
          "amountOfWater": {
            "description": "User's daily water consumption in liters",
            "type": "number",
            "example": 2.5
          }
        }
      },
      "water": {
        "type": "object",
        "required": ["_id", "userId", "date", "amount"],
        "properties": {
          "_id": {
            "description": "Contact's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "userId": {
            "description": "User's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af8"
          },
          "date": {
            "description": "Date of water consumption",
            "type": "string",
            "format": "date-time",
            "example": "2023-07-20T00:00:00.000Z"
          },
          "amount": {
            "description": "Amount of water consumed",
            "type": "number",
            "example": 1.5
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequest"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}